代码检视意见修改

+++ attention/common/op_kernel/arch35/pse.h
@@ -20,7 +20,7 @@
 #include "util_regbase.h"
 
 namespace regbaseutil {
-enum class PseLayoutTypeEnum: int64_t{
+enum class PseLayoutTypeEnum: uint32_t{
     PSE_S1S2 = 0,
     PSE_1S2 = 1,
     PSE_SLOPE_BN = 2,
